Add cast.

+++ clients/odbc/driver/SQLGetInfo.c
@@ -1557,7 +1557,8 @@ translateInfoType(SQLUSMALLINT InfoType)
 	case SQL_XOPEN_CLI_YEAR:
 		return "SQL_XOPEN_CLI_YEAR";
 	default:
-		snprintf(unknown, sizeof(unknown), "unknown (%u)", InfoType);
+		snprintf(unknown, sizeof(unknown), "unknown (%u)", 
+			 (unsigned int) InfoType);
 		return unknown;
 	}
 }
